Add Plotter class to plot_util module
